+++ src/main/kotlin/com/github/surpsg/diffcoverage/runconfiguration/DiffCoverageRunProfileState.kt
@@ -1,72 +1,86 @@
-package com.github.surpsg.diffcoverage.runconfiguration
+package com.github.surpsg.diffcoverage.actions
 
 import com.github.surpsg.diffcoverage.DiffCoverageBundle
 import com.github.surpsg.diffcoverage.coroutine.BACKGROUND_SCOPE
 import com.github.surpsg.diffcoverage.domain.DiffCoverageConfiguration
 import com.github.surpsg.diffcoverage.properties.NOT_GRADLE_PROJECT
 import com.github.surpsg.diffcoverage.properties.REPORT_LINK
+import com.github.surpsg.diffcoverage.properties.REPORT_LINK_WITH_ERROR
 import com.github.surpsg.diffcoverage.properties.gradle.DIFF_COVERAGE_TASK
 import com.github.surpsg.diffcoverage.properties.gradle.HTML_REPORT_RELATIVE_PATH
 import com.github.surpsg.diffcoverage.services.CoverageVizualizeService
 import com.github.surpsg.diffcoverage.services.gradle.GradleDiffCoveragePluginService
 import com.github.surpsg.diffcoverage.services.gradle.GradleService
 import com.github.surpsg.diffcoverage.services.notifications.BalloonNotificationService
-import com.intellij.execution.ExecutionResult
-import com.intellij.execution.Executor
-import com.intellij.execution.configurations.RunProfileState
-import com.intellij.execution.runners.ProgramRunner
 import com.intellij.notification.NotificationListener
 import com.intellij.notification.NotificationType
+import com.intellij.openapi.actionSystem.AnAction
+import com.intellij.openapi.actionSystem.AnActionEvent
 import com.intellij.openapi.components.service
 import com.intellij.openapi.project.Project
 import kotlinx.coroutines.launch
+import org.jetbrains.plugins.gradle.settings.GradleProjectSettings
+import org.jetbrains.plugins.gradle.settings.GradleSettings
 import java.nio.file.Paths
 
-class DiffCoverageRunProfileState(
-    private val project: Project,
-    private val diffCoverageRunConfiguration: DiffCoverageRunConfiguration
-) : RunProfileState {
+class RunDiffCoverageAction : AnAction() {
 
-    override fun execute(executor: Executor, runner: ProgramRunner<*>): ExecutionResult? {
+    override fun actionPerformed(e: AnActionEvent) {
+        val project = e.project ?: return
+
         if (!project.service<GradleService>().isGradleProject()) {
             project.service<BalloonNotificationService>().notify(
                 notificationType = NotificationType.ERROR,
                 message = DiffCoverageBundle.message(NOT_GRADLE_PROJECT)
             )
-            return null
+            return
         }
 
+        val settings: GradleProjectSettings = GradleSettings.getInstance(project).linkedProjectsSettings.first()
+
         val diffCoveragePluginService = project.service<GradleDiffCoveragePluginService>()
-        val diffCoverageModule = diffCoveragePluginService.lookupDiffCoveragePluginModule() ?: return null
+        val diffCoverageModule = diffCoveragePluginService.lookupDiffCoveragePluginModule(settings.externalProjectPath)
+            ?: return
 
         BACKGROUND_SCOPE.launch {
             diffCoveragePluginService.obtainCacheableDiffCoverageInfo(diffCoverageModule)?.let { coverageInfo ->
-                buildReportByGradleDiffCoveragePlugin(coverageInfo, diffCoverageModule.second)
-                buildReportIde(coverageInfo)
-            }
-        }
-        return null
-    }
-
-    private fun buildReportIde(coverageInfo: DiffCoverageConfiguration) {
-        if (diffCoverageRunConfiguration.buildReportByIde) {
-            project.service<CoverageVizualizeService>().showCoverage(coverageInfo)
-        }
-    }
-
-    private fun buildReportByGradleDiffCoveragePlugin(coverageInfo: DiffCoverageConfiguration, projectPath: String) {
-        if (diffCoverageRunConfiguration.buildReportByGradlePlugin) {
-            project.service<GradleService>().executeGradleTask(DIFF_COVERAGE_TASK, projectPath) {
-                showDiffCoverageReportNotification(coverageInfo, project)
+                launch {
+                    project.service<CoverageVizualizeService>().showCoverage(coverageInfo)
+                }
+                buildCoverageReportByGradlePlugin(project, diffCoverageModule.second, coverageInfo)
             }
         }
     }
 
-    private fun showDiffCoverageReportNotification(diffCoverageInfo: DiffCoverageConfiguration, project: Project) {
+    private suspend fun buildCoverageReportByGradlePlugin(
+        project: Project,
+        modulePath: String,
+        coverageInfo: DiffCoverageConfiguration
+    ) {
+        val successExecute = project.service<GradleService>().suspendableExecuteGradleTask(
+            DIFF_COVERAGE_TASK,
+            modulePath
+        )
+        val messageKeyToNotificationType = if (successExecute) {
+            REPORT_LINK to NotificationType.INFORMATION
+        } else {
+            REPORT_LINK_WITH_ERROR to NotificationType.ERROR
+        }
+        showDiffCoverageReportNotification(project, coverageInfo, messageKeyToNotificationType)
+    }
+
+    private fun showDiffCoverageReportNotification(
+        project: Project,
+        diffCoverageInfo: DiffCoverageConfiguration,
+        messageKeyToNotificationType: Pair<String, NotificationType>
+    ) {
         val reportUrl = Paths.get(diffCoverageInfo.reportsRoot, HTML_REPORT_RELATIVE_PATH).toUri().toString()
         project.service<BalloonNotificationService>().notify(
+            messageKeyToNotificationType.second,
             notificationListener = NotificationListener.URL_OPENING_LISTENER,
-            message = DiffCoverageBundle.message(REPORT_LINK, reportUrl)
+            message = DiffCoverageBundle.message(messageKeyToNotificationType.first, reportUrl)
         )
     }
+
+    override fun isDumbAware(): Boolean = false
 }